modelForm adding successfully

+++ todo_list/models.py
@@ -8,7 +8,7 @@ class List(models.Model):
     )
     item = models.CharField(max_length=100)
     location = models.CharField(max_length=3, choices=LOCATION, default='in')
-    cant = models.FloatField(blank=True, null=True)
+    cant = models.IntegerField(blank=True, null=True)
     description = models.TextField(blank=True, null=True)
     completed = models.BooleanField(default=False)
